# docs/swagger.yaml
basePath: /
definitions:
  models.Song:
    description: Модель песни
    properties:
      group:
        type: string
      song:
        type: string
    type: object
  models.SongResponse:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
  models.VerseResponse:
    properties:
      page:
        type: integer
      perPage:
        type: integer
      totalVerses:
        type: integer
      verses:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: kostya2211@yandex.ru
    name: Konstantin
    url: https://github.com/Spok95
  description: API для управления онлайн-библиотекой песен
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Song Library API
  version: "1.0"
paths:
  /songs:
    get:
      consumes:
      - application/json
      description: Возвращает список всех песен с фильтрацией и пагинацией
      parameters:
      - description: Фильтр по группе
        in: query
        name: group
        type: string
      - description: Фильтр по названию песни
        in: query
        name: song
        type: string
      - description: Фильтр по дате выпуска
        in: query
        name: releaseDate
        type: string
      - default: 10
        description: Количество песен на странице
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение (для пагинации)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SongResponse'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить список песен
      tags:
      - Songs
    post:
      consumes:
      - application/json
      description: Создаёт новую песню и обогащает её данными
      parameters:
      - description: Данные песни
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SongResponse'
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Добавить песню
      tags:
      - Songs
  /songs/{id}:
    delete:
      description: Удаляет песню из базы данных
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Song not found
          schema:
            type: string
      summary: Удалить песню
      tags:
      - Songs
    get:
      description: Возвращает данные о конкретной песне по её ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SongResponse'
        "404":
          description: Song not found
          schema:
            type: string
      summary: Получить песню
      tags:
      - Songs
    put:
      consumes:
      - application/json
      description: Изменяет информацию о песне по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SongResponse'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Song not found
          schema:
            type: string
      summary: Обновить песню
      tags:
      - Songs
  /songs/{id}/verses:
    get:
      description: Возвращает текст песни, разбитый на куплеты (по \n\n) с пагинацией
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Номер страницы (начиная с 1)
        in: query
        name: page
        type: integer
      - default: 1
        description: Количество куплетов на странице
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VerseResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Song not found
          schema:
            type: string
      summary: Получить куплеты песни
      tags:
      - Songs
swagger: "2.0"